pyproject.toml
[tool.poetry]
name = "mybirth"
version = "0.1.0"
description = ""
authors = ["Kirill Grinchuk <kiralgrin@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
telethon = "^1.40.0"
pyyaml = "^6.0.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
